리스폰스 하트비트 메세지 처리에 RequestSaveTime호출하던 버그 수정

+++ ServerManager/ServerManager.cpp
@@ -232,7 +232,7 @@ namespace Manager
 		DWORD offset = 0;
 		DWORD quitEventResult = 0;
 
-		DWORD lastHeartbeatTime = 0;
+		unsigned long long lastHeartbeatTime = 0;
 
 		while (_serverOn)
 		{
